Fix DDP setup (tested). Add documentation

+++ dist_utils.py
@@ -12,6 +12,9 @@ from utils import logger
 def is_distributed_slurm_env() -> bool:
     return "SLURM_PROCID" in os.environ and int(os.environ.get("SLURM_NTASKS", "1")) > 1
 
+def is_distributed_activated() -> bool:
+    return "DISTRIBUTED_RUN" in os.environ
+
 
 def get_rank() -> int:
     if is_distributed_slurm_env():
@@ -30,6 +33,8 @@ def is_rank0() -> bool:
 
 
 def maybe_init_distributed(activate_distributed: bool) -> Tuple[int, int]:
+    if activate_distributed:
+        os.environ["DISTRIBUTED_RUN"] = "1"
     if activate_distributed and is_distributed_slurm_env():
         print("--DETECTED DISTRIBUTED SETUP, DO DISTRIBUTED TRAINING!--")
         import torch.distributed as dist
@@ -53,23 +58,26 @@ def maybe_init_distributed(activate_distributed: bool) -> Tuple[int, int]:
                 f"DDP initialized: world_size={world_size}, local_rank={local_rank}, rank={get_rank()}, pid={os.getpid()}"
             )
         return local_rank, world_size
+    elif activate_distributed and not is_distributed_slurm_env():
+        exit("Try running distributed training but environment is not setup for this!")
     else:
         print("--NOT RUNNING DISTRIBUTED TRAINING!--")
         return 0, 1
 
 
 def maybe_cleanup_distributed():
-    if is_distributed_slurm_env():
+    if is_distributed_activated() and is_distributed_slurm_env():
         import torch.distributed as dist
-
+        # Wait for all processes to be finished and then destroy
+        dist.barrier()
         dist.destroy_process_group()
-        log_rank0("DDP cleaned-up!")
+        print("DDP cleaned-up!")
 
 
 # --- end DDP minimal setup ---
 
 def log_rank(msg, rank):
-    if is_rank_eq(rank):
+    if not is_distributed_activated() or is_rank_eq(rank):
         logger.info(msg)
 
 def log_rank0(msg):
